refactor: 일반 사용자와 관리자의 예약 기능 클래스 분리
