+++ src/main/java/roomescape/controller/AdminReservationController.java
@@ -1,0 +1,39 @@
+package roomescape.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.dto.AdminReservationDetailResponse;
+import roomescape.dto.AdminReservationRequest;
+import roomescape.dto.ReservationResponse;
+import roomescape.service.ReservationService;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+public class AdminReservationController {
+    private final ReservationService reservationService;
+
+    public AdminReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
+
+    @PostMapping("/admin/reservations")
+    public ResponseEntity<ReservationResponse> saveReservationByAdmin(
+            @RequestBody AdminReservationRequest reservationRequest) {
+        ReservationResponse reservationResponse = reservationService.saveByAdmin(reservationRequest);
+        return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.id()))
+                .body(reservationResponse);
+    }
+
+    @GetMapping("/admin/reservations-waiting")
+    public List<AdminReservationDetailResponse> findAllWaitingReservations() {
+        return reservationService.findAllWaitingReservations();
+    }
+
+    @DeleteMapping("/admin/reservations-waiting/{id}")
+    public ResponseEntity<Void> deleteReservationWaitingByAdmin(@PathVariable long id) {
+        reservationService.deleteById(id);
+        return ResponseEntity.noContent().build();
+    }
+}
