+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,30 +1,15 @@
 package roomescape.controller;
 
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import roomescape.domain.LoginMember;
+import roomescape.dto.*;
+import roomescape.service.PaymentService;
+import roomescape.service.ReservationService;
+
 import java.net.URI;
 import java.time.LocalDate;
 import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import roomescape.domain.LoginMember;
-import roomescape.dto.AdminReservationDetailResponse;
-import roomescape.dto.AdminReservationRequest;
-import roomescape.dto.PaymentRequest;
-import roomescape.dto.PaymentResponse;
-import roomescape.dto.ReservationDetailResponse;
-import roomescape.dto.ReservationPaymentRequest;
-import roomescape.dto.ReservationPaymentResponse;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationResponse;
-import roomescape.service.PaymentService;
-import roomescape.service.ReservationService;
 
 @RestController
 public class ReservationController {
@@ -53,14 +38,6 @@ public class ReservationController {
                 .body(savedReservationResponse);
     }
 
-    @PostMapping("/admin/reservations")
-    public ResponseEntity<ReservationResponse> saveReservationByAdmin(
-            @RequestBody AdminReservationRequest reservationRequest) {
-        ReservationResponse reservationResponse = reservationService.saveByAdmin(reservationRequest);
-        return ResponseEntity.created(URI.create("/reservations/" + reservationResponse.id()))
-                .body(reservationResponse);
-    }
-
     @GetMapping("/reservations")
     public List<ReservationResponse> findAllReservations() {
         return reservationService.findAllReservations();
@@ -81,11 +58,6 @@ public class ReservationController {
         return reservationService.searchReservation(themeId, memberId, dateFrom, dateTo);
     }
 
-    @GetMapping("/admin/reservations-waiting")
-    public List<AdminReservationDetailResponse> findAllWaitingReservations() {
-        return reservationService.findAllWaitingReservations();
-    }
-
     @DeleteMapping("/reservations/{id}")
     public ResponseEntity<Void> delete(@PathVariable long id) {
         reservationService.deleteById(id);
@@ -95,12 +67,6 @@ public class ReservationController {
     @DeleteMapping("/reservations-waiting/{id}")
     public ResponseEntity<Void> deleteReservationWaiting(@Authenticated LoginMember loginMember, @PathVariable long id) {
         reservationService.deleteByMemberIdAndId(loginMember, id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @DeleteMapping("/admin/reservations-waiting/{id}")
-    public ResponseEntity<Void> deleteReservationWaitingByAdmin(@PathVariable long id) {
-        reservationService.deleteById(id);
         return ResponseEntity.noContent().build();
     }
 }
